Change doubles to floats in DTOs

+++ src/SIL.Machine.WebApi/Models/TranslationResultDto.cs
@@ -5,7 +5,7 @@ namespace SIL.Machine.WebApi.Models
 	public class TranslationResultDto
 	{
 		public IReadOnlyList<string> Target { get; set; }
-		public IReadOnlyList<double> Confidences { get; set; }
+		public IReadOnlyList<float> Confidences { get; set; }
 		public IReadOnlyList<AlignedWordPairDto> Alignment { get; set; }
 	}
 }
