+++ src/SIL.Machine.WebApi/Models/WordGraphArcDto.cs
@@ -6,9 +6,9 @@ namespace SIL.Machine.WebApi.Models
 	{
 		public int PrevState { get; set; }
 		public int NextState { get; set; }
-		public double Score { get; set; }
+		public float Score { get; set; }
 		public IReadOnlyList<string> Words { get; set; }
-		public IReadOnlyList<double> Confidences { get; set; }
+		public IReadOnlyList<float> Confidences { get; set; }
 		public int SourceStartIndex { get; set; }
 		public int SourceEndIndex { get; set; }
 		public bool IsUnknown { get; set; }
